# specs/api/api.yml
openapi: 3.0.0
info:
  title: GrowthLog API
  version: 1.0.0
  description: |
    ## Response Envelope 規則

    | 情境 | HTTP Status | Body 格式 |
    |------|------------|-----------|
    | 成功讀取 | 200 | 直接回傳資料物件，無 success 包裝 |
    | 成功寫入 | 200 | `{"success": true}` |
    | 寵物操作成功 | 200 | `{"success": true, "pet": {...}}` |
    | 任何錯誤 | 對應 4xx | `{"success": false, "error": {"violation_type": "<CODE>"}}` |

    ## violation_type → HTTP Status 對應表

    | violation_type | HTTP Status | 語意 |
    |----------------|-------------|------|
    | `UNAUTHORIZED` | 401 | 未帶 token 或 token 無效 |
    | `FORBIDDEN` | 403 | 有效 token 但無權操作他人資源 |
    | `NOT_FOUND` | 404 | 資源不存在 |
    | `MISSING_REQUIRED_FIELD` | 400 | 必填欄位缺失 |
    | `INVALID_VALUE` | 400 | 欄位值不合法 |
    | `INVALID_STATUS_TRANSITION` | 422 | 狀態機轉換不合法 |
    | `ALREADY_EXISTS` | 422 | 業務規則衝突（如寵物已存在）|

    ## 認證

    受保護端點需在 Header 帶入：Authorization: Bearer <JWT>。
    JWT 由 POST /api/auth/google 取得，以 HS256 + JWT_SECRET 環境變數簽發。
    前端將 token 存入 auth-token cookie，每次請求自動帶入。

    ## 路徑前綴

    後端實際路徑加 /api 前綴（如 /api/goals）。
    前端透過 Next.js proxy 將 /api/v1/* rewrite 至後端 /api/*。
    本文件路徑省略 /api 前綴，以描述 API 的語意介面。
servers:
  - url: http://localhost:8000/api
    description: 本地開發後端
security:
  - bearerAuth: []
paths:
...
  /goals:
    get:
      operationId: listGoals
      summary: 查看學習目標列表
      responses:
        '200':
          description: 回傳學習目標列表（依截止日期升冪排序）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalListResponse'
    post:
      operationId: createGoal
      summary: 建立學習目標
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /goals/{goalId}/status:
    patch:
      operationId: updateGoalStatus
      summary: 更新學習目標狀態
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalStatusRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────
  # Learning Records
  # ─────────────────────────────────────────
...
  /moods:
    get:
      operationId: listMoods
      summary: 查看心情紀錄
      responses:
        '200':
          description: 回傳心情紀錄列表（依 entry_date 遞減排序）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodListResponse'
    post:
      operationId: createMood
      summary: 記錄心情
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMoodRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /moods/{entryId}:
    patch:
      operationId: updateMood
      summary: 編輯心情紀錄
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMoodRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      operationId: deleteMood
      summary: 刪除心情紀錄
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────
  # Gratitude
  # ─────────────────────────────────────────
  /gratitude:
    get:
      operationId: listGratitude
      summary: 查看感謝日記
      responses:
        '200':
          description: 回傳感謝條目列表（依 entry_date 遞減排序）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GratitudeListResponse'
    post:
      operationId: createGratitude
      summary: 新增感謝條目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGratitudeRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /gratitude/{entryId}:
    patch:
      operationId: updateGratitude
      summary: 編輯感謝條目
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGratitudeRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      operationId: deleteGratitude
      summary: 刪除感謝條目
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────
  # Pet
  # ─────────────────────────────────────────
  /pet:
    get:
      operationId: getPet
      summary: 查看電子寵物狀態
      responses:
        '200':
          description: 回傳寵物狀態
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createPet
      summary: 建立電子寵物
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePetRequest'
      responses:
        '200':
          description: 建立成功，回傳初始寵物狀態
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
  /pet/feed:
    post:
      operationId: feedPet
      summary: 餵食電子寵物
      responses:
        '200':
          description: 餵食成功，回傳更新後狀態
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /pet/interact:
    post:
      operationId: interactWithPet
      summary: 與電子寵物互動（撫摸）
      responses:
        '200':
          description: 互動成功，回傳更新後狀態
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

# ─────────────────────────────────────────
# Components
# ─────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    SuccessResponse:
      description: 操作成功（HTTP 200）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResult'
    ErrorResponse:
      description: 操作失敗，回傳對應 4xx HTTP status + success=false body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      description: |
        所有錯誤回傳格式（含 UNAUTHORIZED、FORBIDDEN、NOT_FOUND、INVALID_VALUE 等），
        回傳對應 4xx HTTP status code，body 固定為此結構。
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [violation_type]
          properties:
            violation_type:
              type: string
              description: >
                錯誤代碼：UNAUTHORIZED | NOT_FOUND | INVALID_VALUE |
                PET_ALREADY_EXISTS | MISSING_REQUIRED_FIELD
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]

    # ── Auth ────────────────────────────────
...
    Goal:
      type: object
      required: [goal_id, title, status]
      properties:
        goal_id:
          type: integer
        title:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [not_started, in_progress, completed, abandoned]
    GoalListResponse:
      type: object
      required: [goals]
      properties:
        goals:
          type: array
          items:
            $ref: '#/components/schemas/Goal'
    CreateGoalRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    UpdateGoalStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [not_started, in_progress, completed, abandoned]

    # ── Records ─────────────────────────────
...
    MoodEntry:
      type: object
      required: [entry_id, mood_type, entry_date]
      properties:
        entry_id:
          type: integer
        mood_type:
          type: string
          enum: [depressed, sad, unhappy, neutral, good, happy]
        note:
          type: string
        entry_date:
          type: string
          format: date
    MoodListResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/MoodEntry'
    CreateMoodRequest:
      type: object
      required: [mood_type]
      properties:
        mood_type:
          type: string
          enum: [depressed, sad, unhappy, neutral, good, happy]
        note:
          type: string
        tag_ids:
          type: array
          items:
            type: integer
    UpdateMoodRequest:
      type: object
      required: [mood_type]
      properties:
        mood_type:
          type: string
          enum: [depressed, sad, unhappy, neutral, good, happy]
        note:
          type: string

    # ── Gratitude ───────────────────────────
    GratitudeEntry:
      type: object
      required: [entry_id, content, entry_date]
      properties:
        entry_id:
          type: integer
        content:
          type: string
        entry_date:
          type: string
          format: date
    GratitudeListResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/GratitudeEntry'
    CreateGratitudeRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
    UpdateGratitudeRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── Pet ─────────────────────────────────
    Pet:
      type: object
      required: [species, pet_name, happiness, fullness, level]
      properties:
        species:
          type: string
          enum: [capybara, dog, cat, snake, rabbit, hamster, panda, penguin, fox, dragon]
        pet_name:
          type: string
        happiness:
          type: integer
        fullness:
          type: integer
        level:
          type: integer
    PetResponse:
      type: object
      required: [pet]
      properties:
        pet:
          $ref: '#/components/schemas/Pet'
    CreatePetRequest:
      type: object
      required: [species, pet_name]
      properties:
        species:
          type: string
          enum: [capybara, dog, cat, snake, rabbit, hamster, panda, penguin, fox, dragon]
        pet_name:
          type: string